@@ -13,34 +13,127 @@
         {
             "type": "composer",
             "url": "https://packages.drupal.org/8"
-        }
+        },
+        {
+          "type": "package",
+          "package": {
+            "name": "ckeditor/inserthtml4x",
+            "version": "2.0.0",
+            "type": "drupal-library",
+            "extra": {
+              "installer-name": "ckeditor/plugins/inserthtml4x"
+            },
+            "dist": {
+              "url": "https://download.ckeditor.com/inserthtml4x/releases/inserthtml4x_2.0_0.zip",
+              "type": "zip"
+            }
+          }
+        },
+        {
+              "type": "package",
+              "package": {
+                "name": "ckeditor-plugin/uploadimage",
+                "version": "4.8.0",
+                "type": "drupal-library",
+                "extra": {
+                  "installer-name": "uploadimage"
+                },
+                "dist": {
+                  "url": "https://download.ckeditor.com/uploadimage/releases/uploadimage_4.8.0.zip",
+                  "type": "zip"
+                    }
+                  }
+                },
+                {
+                  "type": "package",
+                  "package": {
+                    "name": "ckeditor-plugin/uploadwidget",
+                    "version": "4.8.0",
+                    "type": "drupal-library",
+                    "extra": {
+                      "installer-name": "uploadwidget"
+                    },
+                    "dist": {
+                      "url": "https://download.ckeditor.com/uploadwidget/releases/uploadwidget_4.8.0.zip",
+                      "type": "zip"
+                    }
+                  }
+                },
+                {
+                  "type": "package",
+                  "package": {
+                    "name": "ckeditor-plugin/filetools",
+                    "version": "4.8.0",
+                    "type": "drupal-library",
+                    "extra": {
+                      "installer-name": "filetools"
+                    },
+                    "dist": {
+                      "url": "https://download.ckeditor.com/filetools/releases/filetools_4.8.0.zip",
+                      "type": "zip"
+                    }
+                  }
+                },
+                {
+                  "type": "package",
+                  "package": {
+                    "name": "ckeditor-plugin/notification",
+                    "version": "4.8.0",
+                    "type": "drupal-library",
+                    "extra": {
+                      "installer-name": "notification"
+                    },
+                    "dist": {
+                      "url": "https://download.ckeditor.com/notification/releases/notification_4.8.0.zip",
+                      "type": "zip"
+                    }
+                  }
+                },
+                {
+                  "type": "package",
+                  "package": {
+                    "name": "ckeditor-plugin/notificationaggregator",
+                    "version": "4.8.0",
+                    "type": "drupal-library",
+                    "extra": {
+                      "installer-name": "notificationaggregator"
+                    },
+                    "dist": {
+                      "url": "https://download.ckeditor.com/notificationaggregator/releases/notificationaggregator_4.8.0.zip",
+                      "type": "zip"
+                    }
+                  }
+                }
     ],
     "require": {
-        "php": ">=7",
+        "php": ">=7.4",
+        "ckeditor-plugin/filetools": "4.8.0",
+        "ckeditor-plugin/notification": "4.8.0",
+        "ckeditor-plugin/notificationaggregator": "4.8.0",
+        "ckeditor-plugin/uploadimage": "4.8.0",
+        "ckeditor-plugin/uploadwidget": "4.8.0",
         "composer/installers": "^1.2",
         "cweagans/composer-patches": "^1.6.5",
-        "drupal/addtoany": "^1.15",
+        "drupal/addtoany": "^2.0",
         "drupal/anchor_link": "^2.5",
-        "drupal/better_exposed_filters": "^3.0",
+        "drupal/better_exposed_filters": "^6.0",
         "drupal/bootstrap_barrio": "^4.22",
-        "drupal/ckeditor_balloonpanel": "^1.1",
+        "drupal/ckeditor_balloonpanel": "^2.0@alpha",
         "drupal/ckeditor_font": "^1.2",
         "drupal/ckeditor_iframe": "^2.1",
-        "drupal/ckeditor_inserthtml": "^1.0",
-        "drupal/ckeditor_smiley": "^1.0",
+        "drupal/ckeditor_smiley": "^1.2",
         "drupal/ckeditor_uploadimage": "^2.0",
-        "drupal/ckeditor_videodetector": "^1.0",
+        "drupal/ckeditor_videodetector": "^2.1",
         "drupal/ckeditor_youtube": "^1.3",
         "drupal/color_field": "^2.4",
         "drupal/colorbutton": "^1.3",
-        "drupal/conditional_fields": "^1.0",
-        "drupal/console": "^1.0.2",
+        "drupal/conditional_fields": "^4.0@alpha",
         "drupal/contact_storage": "^1.0",
-        "drupal/core": "^8.6.0",
         "drupal/core-composer-scaffold": "^9.4",
+        "drupal/core-recommended": "^9.4",
         "drupal/country": "^1.0@RC",
         "drupal/ctools": "^3.2",
-        "drupal/devel": "^2.0",
+        "drupal/devel": "^5.1",
         "drupal/ds": "^3.3",
         "drupal/facets": "^1.4",
         "drupal/flags": "^1.0@beta",
