@@ -4,8 +4,7 @@
  * @file
  * Includes the autoloader created by Composer.
  *
- * This file was generated by drupal-composer/drupal-scaffold.
- * https://github.com/drupal-composer/drupal-scaffold
+ * This file was generated by drupal-scaffold.
  *
  * @see composer.json
  * @see index.php
