@@ -1,9 +1,9 @@
-(function ($, Drupal, drupalSettings) {
+(function($, Drupal, drupalSettings) {
   Drupal.behaviors.loadMap = {
-    attach: function (context, settings) {
+    attach: function(context, settings) {
       $("#gbv-country-map", context)
         .once("loadMap")
-        .each(function () {
+        .each(function() {
           mapboxgl.accessToken = drupalSettings.accessToken;
           let map = new mapboxgl.Map({
             container: "gbv-country-map",
@@ -24,18 +24,19 @@
           map.doubleClickZoom.disable();
           map.addControl(new mapboxgl.NavigationControl());
           let mapData = JSON.parse(drupalSettings.mapData);
-          $(document).ready(function () {
+          $(document).ready(function() {
             let hr_info = mapData.hr_info ? '<a href="' + mapData.hr_info + '" target="_blank">HR Info</a>' : '';
             let countryDescription = mapData.description ? '<p>' + mapData.description + '</p>' : '';
             $(
               '<div class="gbv-country-content-cont"><div class="gbv-country-content"><h3>' +
-                mapData.name +
-                '</h3>' +
-                countryDescription +
-                hr_info +
+              mapData.name +
+              '</h3>' +
+              countryDescription +
+              hr_info +
               '</div></div>'
             ).appendTo("#gbv-global-map");
           });
+          let countryColor = mapData["color-code"] ?? '#A991B6';
           let filteredCountries = {
             type: "FeatureCollection",
             features: [],
@@ -44,13 +45,13 @@
             url: "https://gistcdn.githack.com/timilsinabishal/1df12bb0ce3afe5dbdd6081f89513cae/raw/ece2a05253c7e86a43d8f3e5ea4841cd79b59299/world-admin-0.geojson",
             dataType: "json",
           });
-            map.on("load", function () {
-          postResult.then(function (countries) {
-            let filteredCountry = countries.features.filter(function (feature) {
-              return feature.iso_2 === mapData["iso-2"];
-            });
+          map.on("load", function() {
+            postResult.then(function(countries) {
+              let filteredCountry = countries.features.filter(function(feature) {
+                return feature.iso_2 === mapData["iso-2"];
+              });
 
-            filteredCountries.features.push(filteredCountry[0]);
+              filteredCountries.features.push(filteredCountry[0]);
 
               map.addLayer({
                 id: "country-boundaries",
@@ -61,7 +62,7 @@
                 "source-layer": "country_boundaries",
                 type: "fill",
                 paint: {
-                  "fill-color": '#A991B6',
+                  "fill-color": countryColor,
                   "fill-opacity": 0.7,
                 },
               });
